- fixed memory[free] for HP-UX when >4GB of memory (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@1707 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/sysinfo.c
@@ -1883,7 +1883,7 @@ int	FREEMEM(const char *cmd, const char *parameter,double  *value)
 */
 		/* Free memory in bytes */
 
-			*value=dyn.psd_free * page;
+			*value=(double)dyn.psd_free * page;
 			return SYSINFO_RET_OK;
 		}
 	}
